move more items from common to core project
