+++ src/WeihanLi.Core/Compressor/DataCompressor.cs
