+++ src/WeihanLi.Core/Helpers/Combinatorics/Permutations.cs
