+++ src/WeihanLi.Core/Helpers/PeriodBatching/BatchedConnectionStatus.cs
