+++ src/WeihanLi.Core/Helpers/PeriodBatching/PeriodicBatching.cs
