+++ src/WeihanLi.Core/Helpers/PeriodBatching/PortableTimer.cs
